fix(sglang): Remove unused scheduler_metrics_collector
