+++ rlinf/workers/rollout/sglang/sglang_worker.py
@@ -14,10 +14,7 @@
 
 import asyncio
 import dataclasses
-import json
-import os
 import time
-from dataclasses import dataclass
 from typing import Dict, List
 
 import torch
@@ -37,135 +34,15 @@ from rlinf.scheduler.dynamic_scheduler.manager import RolloutScalingScheduler
 from rlinf.scheduler.dynamic_scheduler.utils import (
     get_scheduler_channel,
 )
-from rlinf.utils.placement import (
-    ModelParallelComponentPlacement,
-    PlacementMode,
-)
+from rlinf.utils.placement import ModelParallelComponentPlacement
 from rlinf.workers.rollout.sglang import Engine, io_struct
 from rlinf.workers.rollout.utils import (
+    MetaInfoStatsCollector,
+    RolloutEngineStats,
     RunningStatusManager,
     print_sglang_outputs,
 )
 from toolkits.math_verifier.verify import MathRewardModel, math_verify_call
-
-
-class MetaInfoStatsCollector:
-    """Collector for SGLang meta_info statistics
-
-    This collector is only initialized when enabled via configuration.
-    Add the following parameters to your generation config section:
-
-    generation:
-      collect_meta_stats: true  # Enable meta_info statistics collection
-      meta_stats_file: "custom_meta_stats.jsonl"  # Optional: custom output file
-      async_meta_stats_file: "custom_async_meta_stats.jsonl"  # Optional: custom async output file
-      schedule_meta_stats_file: "custom_schedule_meta_stats.jsonl"  # Optional: custom schedule output file
-    """
-
-    def __init__(self, output_file: str = "sglang_meta_stats.jsonl"):
-        self.output_file = output_file
-        self.stats_buffer = []
-        self.buffer_size = 100  # Write to file every 100 records
-
-        # Ensure output directory exists
-        os.makedirs(
-            os.path.dirname(self.output_file)
-            if os.path.dirname(self.output_file)
-            else ".",
-            exist_ok=True,
-        )
-
-        # Initialize file with header if it doesn't exist
-        if not os.path.exists(self.output_file):
-            with open(self.output_file, "w") as f:
-                f.write("")  # Create empty file
-
-    def collect_batch_stats(self, outputs: List[Dict], batch_id: int) -> None:
-        """Collect statistics from a batch of SGLang outputs
-
-        Args:
-            outputs: List of SGLang output dictionaries
-            batch_id: Unique identifier for this batch
-        """
-        current_time = time.time()
-
-        for req_idx, output in enumerate(outputs):
-            try:
-                # Extract meta_info
-                meta_info = output.get("meta_info", {})
-
-                # Extract the specific metrics you requested
-                stats_record = {
-                    "timestamp": current_time,
-                    "batch_id": batch_id,
-                    "request_id": f"batch_{batch_id}_req_{req_idx}",
-                    "prompt_tokens": meta_info.get("prompt_tokens", None),
-                    "completion_tokens": meta_info.get("completion_tokens", None),
-                    "e2e_latency": meta_info.get("e2e_latency", None),
-                    "ttft": meta_info.get("ttft", None),
-                    # Additional useful meta_info fields (if available)
-                    "finish_reason": meta_info.get("finish_reason", {}).get(
-                        "type", None
-                    ),
-                    "total_tokens": (
-                        meta_info.get("prompt_tokens", 0)
-                        + meta_info.get("completion_tokens", 0)
-                    )
-                    if meta_info.get("prompt_tokens") is not None
-                    and meta_info.get("completion_tokens") is not None
-                    else None,
-                    # Add any other meta_info fields that might be useful
-                    "meta_info_keys": list(
-                        meta_info.keys()
-                    ),  # For debugging/inspection
-                }
-
-                self.stats_buffer.append(stats_record)
-
-            except Exception as e:
-                # Log error but continue processing
-                error_record = {
-                    "timestamp": current_time,
-                    "batch_id": batch_id,
-                    "request_id": f"batch_{batch_id}_req_{req_idx}",
-                    "error": str(e),
-                    "output_keys": list(output.keys())
-                    if isinstance(output, dict)
-                    else "not_dict",
-                }
-                self.stats_buffer.append(error_record)
-
-        # Write to file if buffer is full
-        if len(self.stats_buffer) >= self.buffer_size:
-            self._flush_to_file()
-
-    def _flush_to_file(self) -> None:
-        """Write buffered statistics to file"""
-        if not self.stats_buffer:
-            return
-
-        with open(self.output_file, "a") as f:
-            for record in self.stats_buffer:
-                f.write(json.dumps(record) + "\n")
-
-        print(f"Written {len(self.stats_buffer)} records to {self.output_file}")
-        self.stats_buffer = []
-
-    def finalize(self) -> None:
-        """Flush any remaining data and close"""
-        self._flush_to_file()
-        print(f"Finalized stats collection. Data saved to {self.output_file}")
-
-
-@dataclass
-class SchedulerStats:
-    num_running_reqs: int = 0
-    max_running_reqs: int = 0
-    num_used_tokens: int = 0
-    max_total_num_tokens: int = 0
-    token_usage: float = 0.0
-    gen_throughput: float = 0.0
-    num_queue_reqs: int = 0
 
 
 class SGLangWorker(Worker):
@@ -203,7 +80,7 @@ class SGLangWorker(Worker):
         self._collect_meta_stats = getattr(
             self._cfg.rollout, "collect_meta_stats", False
         )
-        self._use_auto_scheduler = self._placement._placement_mode == PlacementMode.AUTO
+        self._use_auto_scheduler = self._placement.is_auto
 
         if self._collect_meta_stats:
             self._init_meta_stats_collector()
@@ -230,27 +107,11 @@ class SGLangWorker(Worker):
         self.async_meta_stats_collector = MetaInfoStatsCollector(async_stats_file)
         self.async_batch_counter = 0
 
-        if self._use_auto_scheduler:
-            schedule_stats_file = getattr(
-                self._cfg.rollout,
-                "schedule_meta_stats_file",
-                f"sglang_meta_stats_rank_{self._rank}.jsonl",
-            )
-            self.schedule_meta_stats_collector = MetaInfoStatsCollector(
-                schedule_stats_file
-            )
-            self.schedule_batch_counter = 0
-
     def _collect_stats(self, engine_results: List[Dict]):
-        if self._collect_meta_stats:
-            self.async_meta_stats_collector.collect_batch_stats(
-                engine_results, self.async_batch_counter
-            )
-            self.async_batch_counter += 1
-
-        if self._use_auto_scheduler:
-            # TODO: collect scheduler stats
-            pass
+        self.async_meta_stats_collector.collect_batch_stats(
+            engine_results, self.async_batch_counter
+        )
+        self.async_batch_counter += 1
 
     def _get_sampling_param_from_config(self) -> dict:
         """
@@ -337,11 +198,8 @@ class SGLangWorker(Worker):
         Shutdown the SGLang task.
         """
         # Finalize meta_info statistics collectors if they exist
-        if hasattr(self, "async_meta_stats_collector"):
+        if self._collect_meta_stats:
             self.async_meta_stats_collector.finalize()
-
-        if hasattr(self, "schedule_meta_stats_collector"):
-            self.schedule_meta_stats_collector.finalize()
 
         self.log_info(f"Shutting down SGLang worker {self._rank} ...")
         self._engine.shutdown()
@@ -453,7 +311,7 @@ class SGLangWorker(Worker):
         state = await self._engine.tokenizer_manager.run_task_method(
             io_struct.TaskMethodInput(method_name="get_scheduler_running_state")
         )
-        state = SchedulerStats(**state)
+        state = RolloutEngineStats(**state)
 
         return state
 
@@ -506,7 +364,7 @@ class SGLangWorker(Worker):
         return seq_group_info
 
     async def rollout(self, input_channel: Channel, output_channel: Channel):
-        if self.get_parent_rank() == 0:
+        if self._rank == 0:
             self.log_info("Starting async generation...")
         request: RolloutRequest = input_channel.get()
         groups = request.to_seq_group_infos()
